// types/character.ts
export type Character = {
  id: string;
  number: string;
  name: string;
  firstName: string;
  lastName: string;
  role: string;
  subtitle: string;
  image: string;
  backgroundImage: string;
  bounty: string;
  devilFruit: string;
  dream: string;
  description: string;
  quote: string;
  story: string;

  birthday: string;
  height: string;
  origin: string;
  affiliation: string;

  theme: {
    text: string;
    gradient: string;
    glow: string;
    border: string;
    line: string;
    color: string;
  };
};